\name{CascadeData-package}
\alias{CascadeData-package}
\alias{CascadeData}
\docType{package}
\title{The Data for the Cascade Package}
\description{
Datasets used as examples in the vignette of the Cascade package, a modeling tool allowing gene selection, reverse engineering, and prediction in Cascade networks.
}
\details{
\tabular{ll}{
Package: \tab CascadeData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2019-01-27\cr
License: \tab GNU 2.0\cr
}

}
\author{
This package has been written by Fr\'ed\'eric Bertrand, Myriam Maumy-Bertrand and Nicolas Jung with biological insights from Laurent Vallat.
Maintainer:  Fr\'ed\'eric Bertrand <frederic.bertrand@math.unistra.fr>
}


\references{
Jung, N., Bertrand, F., Bahram, S., Vallat, L., and Maumy-Bertrand, M. (2013). Cascade: a R-package to study, predict and simulate the diffusion of a signal through a temporal gene network. \emph{Bioinformatics}, btt705. 


	
Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., ... & Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. \emph{Proceedings of the National Academy of Sciences}, 110(2), 459-464.
}

\keyword{ package }

