\name{micro_S}
\alias{micro_S}
\docType{data}
\title{
Stimulated data
}
\description{
Stimulated data
}
\usage{data(micro_S)}
\format{
  A data frame with 54613 probesets measured 6 times throught 4 time points.
}

\references{
Vallat, L., Park, Y., Li, C., & Gribben, J. G. (2007). Temporal genetic program following B-cell receptor cross-linking: altered balance between proliferation and death in healthy and malignant B cells. Blood, 109(9), 3989-3997.


Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., ... & Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. Proceedings of the National Academy of Sciences, 110(2), 459-464.}

\examples{
data(micro_S)
}
\keyword{datasets}
