\name{micro_S}
\alias{micro_S}
\docType{data}
\title{
Stimulated data
}
\description{
This is the stimulated data part of the GSE39411 dataset, \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE39411}. Data were normalized and are ready to use.
}
\usage{data(micro_S)}
\format{
  A data frame with 54613 probesets measured 6 times throught 4 time points.
}

\references{
Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., \dots, Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. Proceedings of the National Academy of Sciences, 110(2), 459-464, \url{https://dx.doi.org/10.1073/pnas.1211130110}.}

\examples{
data(micro_S)
}
\keyword{datasets}
